Add Recipes.js and TotalCountry.js components
