+++ src/components/Country.js
@@ -1,11 +1,12 @@
 import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
 import ShowIcon from '../assets/images/ShowIcon.png';
 
 const Country = (props) => {
   const { idCountry, NameCountry, ImageCountry } = props;
   return (
     <li className="country-container" key={idCountry}>
-      <button type="button" className="show-icon"><img src={ShowIcon} alt="Show icon" /></button>
+      <button type="button" className="show-icon"><Link to="/list"><img src={ShowIcon} alt="Show icon" /></Link></button>
       <img src={ImageCountry} alt="Country" />
       <h2 className="country-text">{NameCountry}</h2>
       <p className="text-views">954</p>
