+++ src/components/Header.js
@@ -1,3 +1,4 @@
+import { Link } from 'react-router-dom';
 import GoBackIcon from '../assets/images/GoBackIcon.png';
 import MicIcon from '../assets/images/MicIcon.png';
 import SettingsIcon from '../assets/images/SettingsIcon.png';
@@ -5,7 +6,7 @@ import SettingsIcon from '../assets/images/SettingsIcon.png';
 const Header = () => (
   <header>
     <nav>
-      <button type="button" className="navigation-icons"><img src={GoBackIcon} alt="Go back icon" /></button>
+      <button type="button" className="navigation-icons"><Link to="/"><img src={GoBackIcon} alt="Go back icon" /></Link></button>
       <h1>I Love Food</h1>
       <div>
         <button type="button" className="navigation-icons"><img src={MicIcon} alt="Go back icon" /></button>
