+++ src/components/TotalCountry.js
@@ -1,0 +1,13 @@
+import Country11 from '../assets/images/country-1-1.png';
+
+const TotalCountry = () => (
+  <section className="total-country">
+    <img className="world-map-img" src={Country11} alt="Country" />
+    <div className="total-world-container-text">
+      <h2 className="name-country">Argentina</h2>
+      <p className="text-views">954</p>
+    </div>
+  </section>
+);
+
+export default TotalCountry;
